room: fix abyss height check
As Lara's position is stored as int32, we need to use the same in the
abyss height check function, otherwise internal casting yields invalid
results.

Resolves #2874.

+++ src/libtrx/game/rooms/common.c
@@ -550,7 +550,7 @@ void Room_SetAbyssHeight(const int16_t height)
     CLAMPG(m_AbyssMaxHeight, MAX_HEIGHT - STEP_L);
 }
 
-bool Room_IsAbyssHeight(const int16_t height)
+bool Room_IsAbyssHeight(const int32_t height)
 {
     return m_AbyssMinHeight != 0 && height >= m_AbyssMinHeight;
 }
