+++ src/libtrx/include/libtrx/game/rooms/common.h
@@ -43,7 +43,7 @@ SECTOR *Room_GetPitSector(const SECTOR *sector, int32_t x, int32_t z);
 SECTOR *Room_GetSkySector(const SECTOR *sector, int32_t x, int32_t z);
 
 void Room_SetAbyssHeight(int16_t height);
-bool Room_IsAbyssHeight(int16_t height);
+bool Room_IsAbyssHeight(int32_t height);
 HEIGHT_TYPE Room_GetHeightType(void);
 int16_t Room_GetHeight(const SECTOR *sector, int32_t x, int32_t y, int32_t z);
 int16_t Room_GetHeightEx(
